Fix feign in cart service

+++ commerce/interaction-api/src/main/java/ru/yandex/practicum/interactionapi/request/AddProductToWarehouseRequest.java
@@ -14,6 +14,7 @@ import java.util.UUID;
 public class AddProductToWarehouseRequest {
     @NotNull
     UUID productId;
+    
     @NotNull
-    Long quantity;
+    int quantity;
 }
